+++ test/repository/institutions/UniversityRepositoryTest.scala
@@ -1,4 +1,4 @@
-package repositories.institutions
+package repository.institutions
 
 import domain.institutions.University
 import org.scalatest.FunSuite
