+++ test/repository/location/LocationRepositoryTest.scala
@@ -1,4 +1,4 @@
-package repositories.location
+package repository.location
 
 import domain.location.Location
 import org.scalatest.FunSuite
