+++ test/repository/users/UserPasswordRepositoryTest.scala
@@ -1,4 +1,4 @@
-package repositories.users
+package repository.users
 
 import domain.users.UserPassword
 import org.scalatest.FunSuite
