+++ test/repository/users/UserRepositoryTest.scala
@@ -1,4 +1,4 @@
-package repositories.users
+package repository.users
 
 import java.time.LocalDateTime
